ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
scikit-image==0.23.2
python-multipart==0.0.9
pydantic==2.7.1
